pyproject.toml
[project]
name = "aion2c"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
